fix signals
